feat: Allow to pass entry point to clipstick parse function
This is helpful, when list of CLI args comes not from `sys.argv` and,
as result, entry point cannot be parsed from `sys.argv[0]`.

This is also might be helpful for cases, when some CLI called via
`python3 -m ...` and `sys.argv[0]` as result contain meaningless value
of `something/__main__.py` (however, looks like, the proper fix for that
case is out of scope of current change).